+++ HISTORY.rst
@@ -5,6 +5,9 @@ History
 1.0dev
 ------
 
+- add display renderer
+  [rnix, 2011-08-03]
+
 - adopt to yafowil 1.1
   [rnix, 2011-07-08]